Fix naming of module
